+++ src/bullkpy/pl/__init__.py
@@ -1,5 +1,22 @@
 from ._style import set_style, _savefig, get_palette
- 
+
+from ._colors import _get_series, get_categorical_colors, categorical_colors_array
+
+from .association import association_heatmap, boxplot_with_stats, categorical_confusion
+from .association_rankplots import (
+    rankplot_association, dotplot_association, heatmap_association,
+    )
+from .clustering_plots import ari_resolution_heatmap, categorical_confusion
+from .corr_heatmap import corr_heatmap, gene_panel_correlation_heatmap
+from .corrplot import corrplot
+from .de import volcano, rankplot, ma 
+from .dotplot import dotplot
+from .gene_plot import gene_plot
+from .heatmap_de import heatmap_de
+
+from .pca import pca_scatter, pca_variance_ratio
+from .pca_loadings import pca_loadings_bar, pca_loadings_heatmap
+
 from .qc import (
     qc_metrics,
     library_size_vs_genes,
@@ -8,25 +25,11 @@ from .qc import (
     qc_scatter_panel,
     qc_by_group, qc_pairplot,
 )
-
-from .pca import pca_scatter, pca_variance_ratio
-from .umap import umap
-from .de import volcano, rankplot, ma
-from .dotplot import dotplot
-from .heatmap_de import heatmap_de
-from .violin import violin
-from .corr_heatmap import corr_heatmap, gene_panel_correlation_heatmap
-from .pca_loadings import pca_loadings_bar, pca_loadings_heatmap
-from .gene_plot import gene_plot
 from .rank_genes_groups import rank_genes_groups, rank_genes_groups_dotplot
 from .sample_distances import sample_distances, sample_correlation_clustergram
-from ._colors import _get_series, get_categorical_colors, categorical_colors_array
-from .clustering_plots import ari_resolution_heatmap, categorical_confusion
-from .corrplot_obs import corrplot_obs
-from .association import association_heatmap, boxplot_with_stats, categorical_confusion
-from .association_rankplots import (
-    rankplot_association, dotplot_association, heatmap_association,
-    )
+from .umap import umap
+from .violin import violin
+
 
 from .dispersion import (
     axis_vs_dispersion, dispersion_summary,
@@ -61,7 +64,8 @@ from .oncoprint import oncoprint
 
 from .gsea_bubbleplot import gsea_bubbleplot
 
-__all__ = ["qc_metrics", "library_size_vs_genes",
+__all__ = ["corrplot",
+           "qc_metrics", "library_size_vs_genes",
            "mt_fraction_vs_counts", "genes_vs_mt_fraction",
            "qc_scatter_panel",
            "set_style", "_savefig", "get_palette",
@@ -79,7 +83,7 @@ __all__ = ["qc_metrics", "library_size_vs_genes",
            "sample_distances", "sample_correlation_clustergram",
            "_get_series", "get_categorical_colors", "categorical_colors_array",
            "ari_resolution_heatmap", "categorical_confusion",
-           "corrplot_obs",
+           
            "association_heatmap", "boxplot_with_stats", "categorical_confusion",
            "rankplot_association", "dotplot_association", "heatmap_association",
            "gene_association", "volcano_categorical", "gene_association_volcano",